+++ waspc/ChangeLog.md
@@ -4,6 +4,7 @@
 
 ### 🎉 New Features
 
+- You can now specify which PostgreSQL image to use in `wasp start db` with the `--db-image` argument. ([#3182](https://github.com/wasp-lang/wasp/pull/3182))
 - You can now specify which PostgreSQL image to use in `wasp deploy railway` with the `--db-image` argument. ([#3184](https://github.com/wasp-lang/wasp/pull/3184))
 
 ### 📖 Documentation